+++ src/schema.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use serde::Serialize;
 
 /// A schema type representing a variably encoded integer